+++ CHANGELOG.md
@@ -3,6 +3,11 @@ All notable changes to this package will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
+## [7.10.0-exp.1] - 2022-05-27
+- Added new image assert function that implements functionality to do image asserts with floating point images. These images can have negative numbers and are written out as EXR.
+- Added new image comparison function that implements MSE/RMSE comparisons.
+- Added support for applying texture import settings for the written ActualImages.
+
 ## [7.9.0-exp.1] - 2022-05-18
 - Added SRP Test Scene Asset. Which allows users to add scenes combined with SRP Assets. So that we can test 1 scene with several different SRP's.